修改flyway脚本

+++ src/main/resources/db/migration/V13__Create_soup_table.sql
@@ -1,7 +1,7 @@
 create table soup
 (
     id int,
-    gmt_update int not null,
+    gmt_update bigint not null,
     content varchar(200) not null,
     constraint soup_pk
         primary key (id)
